requirements.txt
selenium
webdriver-manager
pytesseract
Pillow
